+++ src/data/attributeState.js
@@ -1,0 +1,34 @@
+const attributeState = [
+  {
+    type: "set_accousticness",
+    name: "Accousticness",
+    value: 0.5,
+  },
+  {
+    type: "set_danceability",
+    name: "Danceability",
+    value: 0.5,
+  },
+  {
+    type: "set_energy",
+    name: "Energy",
+    value: 0.5,
+  },
+  {
+    type: "set_instrumentalness",
+    name: "Instrumentalness",
+    value: 0.5,
+  },
+  {
+    type: "set_loudness",
+    name: "Loudness",
+    value: 0.5,
+  },
+  {
+    type: "set_tempo",
+    name: "Tempo",
+    value: 0.5,
+  },
+];
+
+module.exports = { attributeState };
