+++ src/pages/Splash/Splash.jsx
